associate more models
