+++ rust/src/tool-parser/src/kimi_k2.rs
@@ -8,7 +8,7 @@ use winnow::token::{literal, rest, take_until, take_while};
 
 use super::utils::{JsonObjectScanState, parse_buffered_event, safe_text_len, take_json_object};
 use super::{Result, ToolCallDelta, ToolParser, ToolParserOutput};
-use crate::Tool;
+use crate::{StructuralTagModel, Tool};
 
 const TOOL_CALLS_START: &str = "<|tool_calls_section_begin|>";
 const TOOL_CALLS_END: &str = "<|tool_calls_section_end|>";
@@ -145,6 +145,10 @@ impl ToolParser for KimiK2ToolParser {
 
     fn preserve_special_tokens(&self) -> bool {
         true
+    }
+
+    fn structural_tag_model(&self) -> Option<StructuralTagModel> {
+        Some(StructuralTagModel::Kimi)
     }
 
     fn tool_call_id(&self, tool_index: usize) -> Option<&str> {
